Fix PIO Linter Warning
This fixes the linter warning for Platformio due to module is defined in platformio.ini and also my_user_config.h

If module is not defined in my_user_config.h nor in platformio.ini, sonoff_post.h will make SONOFF_BASIC as the default.

This works for both Arduino and Platformio

This will work as the same as #define MY_LANGUAGE works. If you define it in my_user_config.h, the compiler will use that language, if not defined, the compiler will use en-GB as the default language.