Add Option to Activate PWM CT mode without select module 48
Hello,

I would like to share with you my extension of the code.
I have some new lamps that have 2 PWM channels that are not classically wired. The first one is for brightness and the other channel is for color temperature.
Tasmota has already integrated the possibility to switch between CWWW and CTBRI. I only built the possibility to set it up manually.

greetings,
Jens

PS: This is my first commit. If I did something wrong, please report it.