Sm2135: Fixed power toggling
Condition distinguishing between RGB or CW channel updates in
Sm2135SetChannels() did not work well when power was toggled.

When the device was turned off, RGB values were set to 0, being
interpreted as a CW channel update. Due to that, toggling power
affected only the CW channel and RGB LEDs were always on.